You've got to be kidding me Xcode.

+++ ZuseTests/ZSCodeTraverserTests.m
@@ -211,7 +211,7 @@
     }];
     
     XCTAssertEqual((NSUInteger)0, statements.count, @"");
-    XCTAssertEqual((NSUInteger)7, timesCalled, @"");
+    XCTAssertEqual(7, timesCalled, @"");
 }
 
 - (void)testMapBlockReplacesItems {
